refactor(editor): engine controller 분리
